+++ fastapi_restful/api.py
@@ -1,6 +1,6 @@
-from typing import Type, Optional, Dict, Union
+from typing import Dict, Optional, Type, Union
 
-from fastapi import FastAPI, APIRouter
+from fastapi import APIRouter, FastAPI
 
 from .resource import Resource
 
@@ -10,7 +10,7 @@ class APIMixin:
 
     @staticmethod
     def _strip_prefix(prefix: Optional[str]) -> str:
-        return prefix.strip('/') if prefix is not None else ''
+        return prefix.strip("/") if prefix is not None else ""
 
     @property
     def urls(self) -> dict:
@@ -33,15 +33,17 @@ class APIMixin:
 
         Parameters
         ----------
-        resource: Type of Resource
-        path: Resource path
+        resource
+            Type of Resource
+        path
+            Resource path
 
         Returns
         -------
         None
         """
-        res = resource(path)
-        self.router.include_router(res.router)
+        resource_instance = resource(path)
+        self.router.include_router(resource_instance.router)
 
 
 class APIVersion(APIMixin):
@@ -51,7 +53,7 @@ class APIVersion(APIMixin):
 
     def __init__(self, version_prefix: str):
         prefix = self._strip_prefix(version_prefix)
-        self.router = APIRouter(prefix=f'/{prefix}')
+        self.router = APIRouter(prefix=f"/{prefix}")
 
     def __str__(self) -> str:
         return self.router.prefix
@@ -59,13 +61,10 @@ class APIVersion(APIMixin):
 
 class RestAPI(APIMixin):
     """Main class for create RESTful-API."""
+
     router: Union[APIRouter, FastAPI]
 
-    _fastapi: FastAPI
-    _prefix: str
-    _versions: Dict[str, APIVersion]
-
-    def __init__(self, fastapi_app: FastAPI, prefix: Optional[str] = '/api'):
+    def __init__(self, fastapi_app: FastAPI, prefix: Optional[str] = "/api"):
         """
         Init class.
 
@@ -74,10 +73,10 @@ class RestAPI(APIMixin):
         fastapi_app: instance of FastAPI
         prefix: Default prefix in url path.
         """
-        self._fastapi = fastapi_app
-        self._prefix = prefix
-        self._versions = {}
-        self.__init_router__()
+        self._fastapi: FastAPI = fastapi_app
+        self._prefix: str = prefix
+        self._versions: Dict[str, APIVersion] = {}
+        self.__init_main_router__()
 
     def __getitem__(self, item: str) -> Optional[APIVersion]:
         """
@@ -92,11 +91,11 @@ class RestAPI(APIMixin):
         Instance API or None
         """
         prefix = self._strip_prefix(item)
-        return self._versions.get(f'/{prefix}')
+        return self._versions.get(f"/{prefix}")
 
-    def __init_router__(self) -> None:
+    def __init_main_router__(self) -> None:
         """
-        Init router.
+        Init main router.
 
         Returns
         -------
@@ -104,7 +103,7 @@ class RestAPI(APIMixin):
         """
         prefix = self._strip_prefix(self._prefix)
         if prefix:
-            router = APIRouter(prefix=f'/{prefix}')
+            router = APIRouter(prefix=f"/{prefix}")
         else:
             router = self._fastapi
         self.router = router
@@ -113,25 +112,6 @@ class RestAPI(APIMixin):
     def versions(self) -> Dict[str, APIVersion]:
         """Versions map"""
         return self._versions
-
-    def create_version(self, prefix: str) -> APIVersion:
-        """
-        Create new version API.
-
-        Parameters
-        ----------
-        prefix: Prefix version
-
-        Returns
-        -------
-        APIVersion instance
-        """
-        if prefix in self._versions:
-            raise AssertionError(f'This version is exist: {prefix}')
-        api_spec_ver = APIVersion(prefix)
-        self._versions[prefix] = api_spec_ver
-        self.router.include_router(api_spec_ver.router)
-        return api_spec_ver
 
     def apply(self) -> None:
         """
@@ -145,3 +125,18 @@ class RestAPI(APIMixin):
             self.router.include_router(router.router)
         if isinstance(self.router, APIRouter):
             self._fastapi.include_router(self.router)
+
+    def include_api_version(self, api_version: APIVersion) -> None:
+        """
+        Include API version to main router.
+
+        Parameters
+        ----------
+        api_version
+            Instance of APIVersion with included resources.
+        """
+        prefix = api_version.router.prefix
+        if prefix in self._versions:
+            raise AssertionError(f"This version is exist: {prefix}")
+        self._versions[prefix] = api_version
+        self.router.include_router(api_version.router)