+++ tests/test_api.py
@@ -1,68 +1,72 @@
+from unittest.mock import patch
+
 import pytest
 from fastapi import FastAPI
 
 from fastapi_restful import RestAPI
 
 
-def test_init_rest_api():
+def test__init_rest_api__success():
     app = FastAPI()
     api = RestAPI(fastapi_app=app)
     assert api._fastapi is app
-    assert api.router.prefix == '/api'
+    assert api.router.prefix == "/api"
 
 
-def test_init_rest_api_empty_default_prefix():
+def test__init_rest_api__empty_default_prefix__success():
     app = FastAPI()
     api = RestAPI(fastapi_app=app, prefix=None)
     assert api._fastapi is app
+    assert not hasattr(api.router, "prefix")
 
 
-def test_add_resource(resource_for_test, rest_api_for_test):
-    rest_api_for_test.add_resource(resource_for_test)
+@patch("fastapi_restful.resource.Resource")
+def test__add_resource__success(resource_mock, rest_api_instance):
+    assert rest_api_instance.add_resource(resource_mock) is None
 
 
-def test_create_version(rest_api_for_test):
-    assert rest_api_for_test.versions == {}
-    api = rest_api_for_test.create_version('/v1')
-    assert rest_api_for_test.versions == {'/v1': api}
+def test__get_item__success(rest_api_instance, api_version_instance):
+    assert rest_api_instance.versions == {}
+    prefix = api_version_instance.router.prefix
+    rest_api_instance._versions[prefix] = api_version_instance
+    assert rest_api_instance[prefix] is api_version_instance
 
 
-def test_create_version_failed_is_exist_prefix(rest_api_for_test):
-    version_prefix = '/v1'
-    assert rest_api_for_test.versions == {}
-    api = rest_api_for_test.create_version(version_prefix)
-    assert rest_api_for_test.versions == {version_prefix: api}
-    pytest.raises(AssertionError, rest_api_for_test.create_version, version_prefix)
+def test__str__success(api_version_instance):
+    prefix = api_version_instance.router.prefix
+    assert str(api_version_instance) == prefix
 
 
-def test_get_version_by_prefix(rest_api_for_test):
-    assert rest_api_for_test.versions == {}
-    api = rest_api_for_test.create_version('/v1')
-    assert rest_api_for_test['/v1'] is api
-
-
-def test_str_of_instance(rest_api_for_test):
-    prefix = '/v1'
-    api = rest_api_for_test.create_version(prefix)
-    assert str(api) == prefix
-
-
-def test_url_map(rest_api_for_test, resource_for_test):
+def test__url_map__success(rest_api_instance, resource_type):
     def get(self):
         return {}
-    resource_for_test.get = get
-    resource_prefix = '/test'
-    rest_api_for_test.add_resource(resource_for_test, resource_prefix)
-    base_prefix = rest_api_for_test.router.prefix
-    path = f'{base_prefix}{resource_prefix}'
-    assert rest_api_for_test.urls.get(path) is not None
+
+    resource_type.get = get
+    rest_api_instance.add_resource(resource_type)
+    base_prefix = rest_api_instance.router.prefix
+    path = f"{base_prefix}{resource_type.path}"
+    assert rest_api_instance.urls.get(path) is not None
 
 
-def test_apply(rest_api_for_test, resource_for_test):
+def test__apply__success(rest_api_instance, resource_type, api_version_instance):
     def get(self):
         return {}
-    resource_for_test.get = get
-    v1 = rest_api_for_test.create_version('v1')
-    v1.add_resource(resource_for_test, '/test')
-    rest_api_for_test.apply()
-    assert True
+
+    resource_type.get = get
+    api_version_instance.add_resource(resource_type)
+    rest_api_instance.include_api_version(api_version_instance)
+    assert rest_api_instance.apply() is None
+
+
+def test__call_http_methods__success(resource_type):
+    for method_name in resource_type._HTTP_METHODS:
+        method = getattr(resource_type, method_name)
+        assert method is not None
+        method(resource_type)
+
+
+def test__include_api_version__failed__prefix_is_exist(rest_api_instance, api_version_instance):
+    expected_error_msg = "This version is exist: /v1"
+    rest_api_instance._versions[api_version_instance.router.prefix] = api_version_instance
+    exc = pytest.raises(AssertionError, rest_api_instance.include_api_version, api_version_instance)
+    assert str(exc.value) == expected_error_msg