+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+## 0.4.0
+- Add poetry
+- Add linters, formatters and pre-commit
+- Refactor class APIVersion. 
+
 ## 0.3.0
 - Add decorator route_settings
 - Add method apply for class RestAPI